Implement missing getPartition function (#746)
* Implement missing getPartition function

This was previously generated in the schema but the "pseudo parameters" have to be implemented manually, which partition was not.

- Add the partition DNS Suffix to the output as it's a crucial piece of information for users.
- Implement the partition string property as the identifier.
- Include basic documentation on what the properties will look like.

Matching the design of the classic provider - using DNS instead of URL. The source provider object, and therefore URL, is only an internal construct and not a model derived from AWS directly, so DNS seems to be a better choice for the external interface.

* Generate