Add naming trivia implementation to autonaming process
Introduces the concept of `NamingTrivia` which is a structured set of information
about the semantics of names expected by AWS. Currently, this is most relevant to
FIFO queues in SQS, since they must always be suffixed by `.fifo`.  I expect the
application logic to evolve over time, but since each resource attribute is handled
individually, it may never be the case that we need to compose multiple pieces of trivia.