Improve type generation for CF schemas with AnyOf/OneOf (#986)
Improves type generation for CF schemas that use AnyOf or OneOf, by
addressing schema factoring and refactoring generation to avoid drift
between implemtation for these two fields.

Fixes #777 

## Description
There are essentially 3 small fixes in this change:

1) CloudFormation schemas are based on [Json Schema](json-schema.org)
which supports two fields for declaring variants with multiple arms:
`anyOf` and `oneOf`. They are [specified slightly
differently](https://json-schema.org/understanding-json-schema/reference/combining.html),
but for our purposes, we can map both to the OneOf field of pulumi's
TypeSpec. Unfortunately, we'd lost parity between the generation code
for these two as fixes were only applied to one path or the other. This
change consolidates the generation for AnyOf and OneOf to reduce drift
and bring AnyOf up to parity with OneOf.

2) Json Schema (and CF schemas) also allow [schema
factoring](https://json-schema.org/understanding-json-schema/reference/combining.html#factoring-schemas)
where fields included on the containing schema are applied to every arm
in the anyOf/anyOf variant. This change adds support to correctly
generate types from factored schemas. (This is the root fix for #777.)

3) Often the arms of a oneOf/anyOf don't differ in ways that manifest in
the pulumi type/schema. (eg. two string variants with different formats
as in #777). With this change we deduplicate the types we generate in a
OneOf TypeSpec.

## Testing
There's a severe lack of tests overall in this code, but I added a few
units that are relevant to #777 and carefully sanity checked the changes
to the schema.