RFC: Use devbox to define a consistent dev environment for contributors (#990)
Adds `devbox.json` and `.devcontainer` files to define a reproducible
build environment. Updated the README with instructions for use.

Devbox orchestrates a nix shell with a specified dependency list, which
you can use directly, or inside a container. devcontainer provides a
standard for editors and other tools to work inside in a container.

I personally like this approach as a relatively lightweight way to
maintain a known good build environment; I'm curious to get the team's
feedback.

Ideally, we'd drive our CI off the same docker image for consistency
between local builds and CI (github has some tooling to support this
already: https://github.com/devcontainers/ci); I can add this if we
decide to go down this path.