Mark createOnlyProperties with replaceOnChange (#1053)
Updated gen to add the `repalceOnChanges` flag to any properties
identified in the CF schema as `createOnlyProperties`.

Fixes #736 (mostly; we'll still need to address
https://github.com/pulumi/pulumi-java/issues/1189 to get the correct
behavior in java)

## Details

I've implemented this as a post-processing step on the Resource spec to
avoid cluttering the already complicated type generation and naming
logic.

Ironically, while both the aws CF schema and the sdks use a list of path
strings on the root of the resource to designate
createOnlyProperties/replaceOnChanges, the schema layer uses a flag on
the leaf propertySpecs. So we have to traverse the structure when
setting replaceOnChanges, and then the sdk generation process has to
traverse it again to rebuild the paths.


## Testing
I've added new unit tests and manually verified the schema changes.