Implement skipping credentials validation
We have to put the configuration options in the default provider as these override the stack configuration values, and when attaching the running provider in tests we can't use ENV as the provider is already running with the root `go test` environment.

Fixes #114

Side note - it seems as if the not implemented validation for "skip..." are wrong too as they only trigger if the value is set to a falsy value i.e. don't skip. They don't trigger when "skip" is set to true even though the skip is not actually implemented.