Implement skipping credentials validation (#1326)
Fixes #114 

This allows us to perform previews without needing to have credentials
configured.

Side note - it seems as if the not implemented validation for "skip..."
are wrong too as they only trigger if the value is set to a falsy value
i.e. don't skip. They don't trigger when "skip" is set to true even
though the skip is not actually implemented.

Extracted from #1292