Fix handling of refs to maps
When parsing a property which has a ref to a type which is just a map (using addtionalProperties or patternProperties) we just return an untyped "object".
This adds a special case to check if the referenced type could be represented as a map first and generated any referenced value types.
This fixes 143 resource properties.