Fix parsing of multiple types
In JSSchema the "type" field can either be a string or a list of strings (e.g. "string", "numnber", "object").
Types can also be implied from the other fields set of a schema (e.g. having "Items" indicates it could be an array).

Add handling of where we have multiple types rather than just taking the first type from the array. Model multiple types as a oneOf in the Pulumi schema.

To make a schema easier to work with, we can ensure that all types are explicitly listed in the "type" array via the `NormaliseTypes` function.

This removes 55 instances of Any types and introduces oneOf for these cases so we accurately represent the underlying resource schema.