Fix multi-type resolutions (#1383)
Fixes https://github.com/pulumi/pulumi-aws-native/issues/655
Fixes https://github.com/pulumi/pulumi-aws-native/issues/194

## Fix parsing non-explicitly typed arrays

This changes 6 properties from Any to Array where there was `Items`
defined in the schema, but the type was not explicitly set to `array`.

This also changes one instance of a type from `string` to
`Array<string>` as the type was set to "string" but `Items` was also
defined, so we now prefer the array type.

## Fix parsing of multiple types

In JSSchema the "type" field can either be a string or a list of strings
(e.g. "string", "numnber", "object").
Types can also be implied from the other fields set of a schema (e.g.
having "Items" indicates it could be an array).

Add handling of where we have multiple types rather than just taking the
first type from the array. Model multiple types as a oneOf in the Pulumi
schema.

To make a schema easier to work with, we can ensure that all types are
explicitly listed in the "type" array via the `NormaliseTypes` function.

Where one of the multiple types is "object", we simplify the type to
just be "Any" as this also covers any other primative too.

## SDK Change summary

- `aws-native:apigateway:Method` `requestParameters` additional
properties type changed from `boolean` to `Union<boolean, string>`
- `aws-native:apigatewayv2:Route` `requestParameters` type changed from
`Any` to `Array<{ required: boolean }>`
- Properties changed from `string` to `Any` (becuase it can be
`Union<string, object>`:
  - `aws-native:configuration:ConfigRule` `inputParameters`
  - `aws-native:ec2:VpcEndpoint` `policyDocument`
  - `aws-native:events:Rule` `eventPattern`
- `aws-native:logs:DeliveryDestination` `deliveryDestinationPolicy` type
changed from `Any` to `Array<{ deliveryDestinationName: string,
deliveryDestinationPolicy: string }>`