Add a thin wrapping layer around the CC SDK calls (#1409)
Inspired by the change by @thomas11 in Azure Native, I'm removing the
direct dependency from the provider code to the AWS CloudControl SDK and
adding a thin wrapper layer in-between. I also moved the await logic
into a separate file.

The main purpose of this change is to make the provider code
unit-testable with the CC client mocked out. It's also a first step
towards adding more modularity around the await and error handling
logic.

No functional changes in this PR.