Handle AlreadyExists error during creation (#1439)
The PR does three things:

1. Cover client.Create with unit tests for all branches.
2. Refactor its code to be more concise and to return as quickly as
possible. Return consistent error messages.
3. Add another special case to create logic to process AlreadyExists
error.

Currently, if we receive an AlreadyExists error, we follow our default
flow for all await errors: we proceed to reading the resource state, we
see that it exists, and we return a partial error, which saves this
resource in Pulumi state. This is wrong, because we should not pull an
existing resource into state unexpectedly for a user (without an
import). After the fix, we hard-fail on AlreadyExists.

Fix #1263
Fix #1138