Remove obsolete artifact-cleanup.yml
The action being called is no longer supported, and fails generating P1s for the team.

Despite the file claiming it is auto-generated, the file is not present in ci-mgmt.

Recent versions of GitHub actions automatically expire artifacts after 90 days.

This setting is configurable, so if a more frequent interval is ever needed we can configure the artifact uploading jobs
with an appropriate setting.

In the meanwhile removing this workflow stops the P1s.