chore: augment documentation with CloudFormation docs (#1542)
The CDK team publishes a JSON file with the contents of the
CloudFormation documentation. It can be found here

https://github.com/cdklabs/awscdk-service-spec/tree/main/sources/CloudFormationDocumentation

For whatever reason, the CloudFormation schema does not include a lot of
documentation. This PR uses the CloudFormation docs to augment the
missing documentation. This PR is the first pass at this.

It also includes a report that reports on how many properties have been
augmented and how many still have missing docs. There are still some
that are missing, but I think the number it does populate is pretty good
for a first pass.

```
Number of docs augmented: 14050
Number of docs missing: 818
```

fixes #63, #1336