chore: pin cloudformation docs version
This introduces a mechanism for pinning the version of the
CloudFormation docs that are used. We were just pulling whatever was
latest which sometimes would be different on the next CI run and cause
diffs in the CI job.

Now we will update the docs version at the same time we update the CFN
schema.

closes https://github.com/pulumi/pulumi-aws-native/issues/1605