SDK generation
Had to run the nightly schema generation manually because it was failing
on enums ending in `*Input` & `*Output`. I added a special case for the
couple new ones that came in.

If anyone has any suggestions on a generic solution we could use (e.g.
`Inputx`) I'm open to it.

closes #1640