Use log.Fatal instead of panic and fence the error message
Schema generation emits a lot of warnings and it is hard to see the real blocker error in the noise. Before the change
the blocker error was being sent with `panic`. After the change the code uses `log.Fatal` and adds some `=` fencing to
make it very visually distinct in the output. It is the visually distinct bit that is important.