Refactor Custom Resources to allow configuring timeouts and secrets
Previously it was assumed that the custom resources do not concern
themselves with secretness or handle timeouts.
But this is necessary to support CloudFormation based Custom Resources.
Those can contain secret outputs and support configuring timeouts.

This change modifies the `CustomResource` interface to accomomdate
those requirements. In detail this means that the
Create/Update/Read lifecycle methods now returns a `PropertyMap`
instead of a generic `map[string]interface{}` and take a timeout
parameter.