Extend metadata with descriptors of CF Ref handling (#1836)
CloudFormation Ref intrinsic behavior varies from resource to resource.
This change is an attempt to schematize behavior and expose is in the
provider metadata. One intended use case is `pulumi/pulumi-cdk` CF
emulator.

This change implies the `ref-db.json` becomes the source of truth for
this metadata and we maintain it as well as possible.

About 1/2 of CF resources are currently covered in the file, including
81 of the top used resources
(https://github.com/pulumi/pulumi-cdk/issues/237). Data cleaning and
data label attachment helper code is included but is not intended to be
used in production but only to maintain `ref-db.json`.