Fixes union type conversions to CloudControl (#1850)
This change fixes the conversion of Pulumi input properties to the
format expected by the CloudControl API in the case when Pulumi schema
specifies a union type (OneOf) without a discriminator. Previously the
first matching case would be picked which could erroneously send empty
data to CloudControl. With this change a heuristic is run instead to
pick the non-error case with the largest map or array.

In the long term it would be better to support discriminators and manage
their metadata (#1849).

Fixes #1846