Implement autonaming configuration protocol (#1919)
This PR implement the AWS Native part of
https://github.com/pulumi/pulumi/issues/1518. See
https://github.com/pulumi/pulumi/discussions/17592 for the full design.

In short, https://github.com/pulumi/pulumi/pull/17810 introduced the
protobuf changes required for the provider-side implementation. With
those, a provider can:

- Declare that it supports autonaming configurations with a response
flag in Configure
- Accept two extra properties in CheckRequest: a proposed name and a
mode to apply it

This PR applies those parameters to auto-name calculation if they are
specified:

- In Disabled mode, we don't calculate auto-names and let the resource
validation fail if an explicit name is not specified. For that, I had to
thread an extra `bool` flag through autonaming machinery.
- In Enforce mode, we use whatever ProposedName is supplied by the
engine
- In Propose mode, we take the ProposedName but can also apply trivia
(e.g. add a suffix, as in the e2e test), or check the max/min length

I added unit tests to cover most autonaming cases. Also, added an
end-to-end test in YAML. Apparently, end-2-end tests had no AWS
credentials configured, so I added those. Let me know if that's against
the intention.

Resolves https://github.com/pulumi/pulumi-aws-native/issues/1901