Handle Throttling responses with adaptive retries (#2522)
This fixes what I believe was a significant misconfiguration of the
provider's retry logic.

In particular, we configure our AWS client for 25 retries but this has
no bearing on downstream CloudControl requests. That is, we will retry
when submitting a request to CC, but all of the actual work happens
asynchronously using default retry settings, i.e. 3 retries + ignore
throttling.

Documentation is sparse on CC's error handling, and I had mistakenly
assumed it would inherit the same retry settings we configured our
client with when interacting resources, but this is not the case.

As far as I can tell, the intent (e.g.
https://github.com/pulumi/pulumi-aws-native/pull/862) was to have these
retries apply to the underlying CC Create/Update/Delete operations, but
these always succeed and return a progress handle. We then query that
progress handle, and it eventually indicates the underlying operation
failed with a "Throttling" error code. SDK retries don't kick in because
all of _our_ client's requests -- kicking off the operation, and then
querying for its progress -- were successful.

So CloudControl sees the errors and surfaces them to us as "FAILED"
progress events, but we still want to retry some of the requests,
especially Throttling. I don't want to introduce a bunch of bespoke
retry logic when the AWS SDK already provides a ton of reasonable
defaults, so let's try to leverage that.

This PR makes the following changes:
* We switch from the Standard to the
[AdaptiveMode](https://pkg.go.dev/github.com/aws/aws-sdk-go-v2@v1.39.2/aws/retry#AdaptiveMode)
retryer. AdaptiveMode is essentially the same as Standard, but it will
retry automatically on Throttling errors instead of raising them. This
is important because all of our examples
(https://github.com/pulumi/pulumi-aws-native/issues/1983,
https://github.com/pulumi/pulumi-aws-native/issues/1780,
https://github.com/pulumi/pulumi-cdk/issues/277) are examples of
Throttling errors that our Standard retryer doesn't handle.

> AdaptiveMode provides an experimental retry strategy that expands on
the Standard retry strategy, adding client attempt rate limits. The
attempt rate limit is initially unrestricted, but becomes restricted
when the attempt fails with for a throttle error. When restricted
AdaptiveMode may need to sleep before an attempt is made, if too many
throttles have been received.
  >
> Eventually unrestricted attempt rate limit will be restored once
attempts no longer are failing due to throttle errors.

* We add a method `progressEventToError` which maps error codes on
progress events to first-class error types. This allows us to use our
existing `cfg.Retryer` (the one configured for 25 retries) with these
events.

* We add a `withRetries` method which wraps a Create/Update/Delete
(+await) operation. If during the await we see that the operation failed
due to Throttling, it will be retried using exactly the same retry logic
we've configured our AWS client with.

I was able to reproduce the SQS and GameLift errors and confirmed that
the provider now works as expected with them. Here's an example of an
SQS queue that was created right after being deleted:

```
❯ pulumi up --yes --skip-preview
Updating (dev):
     Type                     Name                 Status            Info
 +   pulumi:pulumi:Stack      aws-native-1959-dev  created (66s)     3 warnings
 +   └─ aws-native:sqs:Queue  foo                  created (65s)

Diagnostics:
  aws-native:sqs:Queue (foo):
```

Along the way the user sees status messages like this:

> Retrying after "Throttling" error (attempt 1/25, waiting 1.579690526s)
> Retrying after "Throttling" error (attempt 2/25, waiting 86.117105ms)
> Retrying after "Throttling" error (attempt 3/25, waiting 4.734073457s)
> Retrying after "Throttling" error (attempt 4/25, waiting 2.157481486s)
>    Retrying after "Throttling" error (attempt 5/25, waiting 20s)

Refs https://github.com/pulumi/pulumi-aws-native/issues/1959.