Handle JSON schema draft-06 exclusiveMaximum in CF schemas
The library we are using for parsing JSON schema (pulumi/jsschema forked
from lestrrat-go/jsschema) is pretty old and does not support the newer
draft-06 JSON schema. The 06 draft changes the schema type of
`exclusiveMaximum` from a boolean to a number.

Long term we should migrate to another library like
[google/jsonschema-go](https://github.com/google/jsonschema-go), but in
the short term i've added a normalization pass over the schema when we
run into this error.